requirements.txt
streamlit
pandas
psycopg2-binary
python-dotenv
bcrypt
